+++ flow-report/src/common.tsx
@@ -6,33 +6,6 @@
 
 import {FunctionComponent} from 'preact';
 import {Util} from '../../report/renderer/util';
-import {SnapshotIcon} from './icons';
-
-const GatherModeIcon: FunctionComponent<{mode: LH.Result.GatherMode}> = ({mode}) => {
-  return (
-    <div
-      className={`GatherModeIcon GatherModeIcon--${mode}`}
-      role="img"
-      aria-label={`Icon representing a ${mode} report`}
-    >
-      {
-        mode === 'snapshot' && <SnapshotIcon/>
-      }
-    </div>
-  );
-};
-
-export const FlowStepIcon: FunctionComponent<{mode?: LH.Result.GatherMode}> = ({mode}) => {
-  return (
-    <div className="FlowStepIcon">
-      <div className="FlowStepIcon__top-line"/>
-      {
-        mode && <GatherModeIcon mode={mode}/>
-      }
-      <div className="FlowStepIcon__bottom-line"/>
-    </div>
-  );
-};
 
 /**
  * Summarizes the category as a ratio of passed audits to total audits.
