+++ flow-report/src/sidebar/flow.tsx
@@ -5,7 +5,7 @@
  */
 
 import {FunctionComponent} from 'preact';
-import {FlowStepIcon} from '../common';
+import {FlowStepIcon} from '../icons';
 import {classNames, useCurrentLhr, useDerivedStepNames, useFlowResult} from '../util';
 import {Separator} from '../common';
